+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - `onIndicatorVisible` callback was added to check when the indicator is visible or not
 - `onValueSelected` was added , it displays the current `double` value selected by the indicator (if the chart is CUSTOM).
 - `onDateTimeSelected` was added, it displays the current `DateTime` selected by the indicator (if the chart is Date type).
+- New date type chart added `BezierChartScale.HOURLY` , doesn't support zoom.
 
 ## [1.0.10]